refactor(rsync_io): split ssh config_lookup into submodules
